Added a WIP screen for FF1 battles and more.
I've been kindn of neglecting GitHub recently; the following is quite a
lot of stuff that should really have been separated.
-Added BattleScreen with basic layout
-Completely overhauled 'Content' class. Now uses an 'Atlas' class based
on TextureAtlases.
-Sprites for battles and world now use texture atlases.
-Added MenuScreen. Currently a temp 3 button thing.
-Moved PlayScreen to the main package.
-Added 'PartyMember' classes. These are implemented by a size 4 array
(as, like in FF1, you always have 4 members). They contain stats of each
character. These should be selected by the player when a new game is
made.